move geomap template filters to a proper template library
